+++ pkg/pub_worker/lib/src/bin/pana_wrapper.dart
@@ -6,7 +6,6 @@ import 'dart:async';
 import 'dart:convert' show json;
 import 'dart:io' show Directory, File, Platform, exit;
 
-import 'package:_pub_shared/search/tags.dart';
 import 'package:logging/logging.dart' show Logger, Level, LogRecord;
 import 'package:pana/pana.dart';
 import 'package:path/path.dart' as p;
@@ -109,11 +108,9 @@ Future<void> main(List<String> args) async {
     package,
     version: version,
     options: InspectOptions(
-      //TODO: Add analysisOptionsYaml, or move the logic into pana
       pubHostedUrl: Platform.environment['PUB_HOSTED_URL']!,
       //TODO: Run dartdoc as part of pana
       checkRemoteRepository: true,
-      futureSdkTag: PackageVersionTags.isDart3Compatible,
     ),
     logger: _log,
     storeResource: (filename, data) async {